+++ stochproc/timeseries/models/__init__.py
@@ -7,6 +7,7 @@ from .ucsv import UCSV
 from .seasonal import Seasonal
 from .smooth_trend import SmoothLinearTrend
 from .trending_ou import TrendingOU
+from .bivariate_ou import BivariateTrendingOU
 
 __all__ = [
     "OrnsteinUhlenbeck",
@@ -17,5 +18,6 @@ __all__ = [
     "UCSV",
     "Seasonal",
     "SmoothLinearTrend",
-    "TrendingOU"
+    "TrendingOU",
+    "BivariateTrendingOU",
 ]